Initial commit
This is my first commit as part of a  project using HTML. I created a file with my favorite
foods, and links to pages with the recipes for those foods. I only used HTML for this
first iteration; I will come back to this project and style it with CSS as I progress in
The Odin Project.

This is part of The Odin Project: Foundations Course (HTML Foundatons Module).